Move vello_api and vello_hybrid to sparse_strips
