+++ sparse_strips/vello_api/src/any.rs
