+++ sparse_strips/vello_api/src/generic_record.rs
