+++ sparse_strips/vello_hybrid/examples/gpu.rs
