+++ sparse_strips/vello_hybrid/examples/simple.rs
