+++ sparse_strips/vello_hybrid/examples/svg.rs
