+++ sparse_strips/vello_hybrid/shader/render.wgsl
