+++ sparse_strips/vello_hybrid/src/fine.rs
