+++ sparse_strips/vello_hybrid/src/flatten.rs
