+++ sparse_strips/vello_hybrid/src/gpu.rs
