+++ sparse_strips/vello_hybrid/src/lib.rs
