+++ sparse_strips/vello_hybrid/src/pixmap.rs
