+++ sparse_strips/vello_hybrid/src/render.rs
