+++ sparse_strips/vello_hybrid/src/simd.rs
