+++ sparse_strips/vello_hybrid/src/strip.rs
