+++ sparse_strips/vello_hybrid/src/tiling.rs
