+++ sparse_strips/vello_hybrid/src/wide_tile.rs
