--- vello_api/README.md
@@ -1,5 +1,0 @@
-# vello_api
-
-Experiment in API abstraction for 2D rendering.
-
-TODO: explain
