--- vello_hybrid/README.md
@@ -1,5 +1,0 @@
-# vello_hybrid
-
-Experiment in hybrid CPU/GPU rendering.
-
-TODO: explain
